phase1 Orderervice

+++ Sizes/DTO/SizeDTO.cs
@@ -10,7 +10,7 @@
 
         public int Quantity { get; set; }
 
-        public decimal StepQuantity { get; set; }
+        public int StepQuantity { get; set; }
 
         public decimal Price { get; set; }
     }
